:sparkles: Modularizando las funciones
