+++ student-blogs/dynamite-4370/helpers/index.js
@@ -1,0 +1,25 @@
+import dayjs from "dayjs";
+import relativeTime from "dayjs/plugin/relativeTime.js";
+await import("dayjs/locale/es-us.js");
+
+dayjs.locale("es-us");
+dayjs.extend(relativeTime);
+
+function humanizeDate(date) {
+    const wrappedDate = dayjs(date);
+    const daysDiff = Math.abs(wrappedDate.diff(Date.now(), "days"));
+
+    const isCurrentYear = wrappedDate.year() === new Date().getFullYear();
+
+    if (!isCurrentYear) {
+        return wrappedDate.format("MMMM DD, YYYY");
+    }else if (daysDiff <= 30){
+        return wrappedDate.fromNow();
+    }else{
+        return wrappedDate.format("MMMM DD");
+    }
+}
+
+export {
+    humanizeDate
+}
